+++ src/SDKs/Compute/Compute.Tests/ScenarioTests/ExtensionTests.cs
@@ -53,7 +53,7 @@ namespace Compute.Tests
                     // Create Storage Account, so that both the VMs can share it
                     var storageAccountOutput = CreateStorageAccount(rgName, storageAccountName);
 
-                    var vm = CreateVM_NoAsyncTracking(rgName, asName, storageAccountOutput, imageRef, out inputVM);
+                    var vm = CreateVM(rgName, asName, storageAccountOutput, imageRef, out inputVM);
 
                     // Delete an extension that does not exist in the VM. A http status code of NoContent should be returned which translates to operation success.
                     m_CrpClient.VirtualMachineExtensions.Delete(rgName, vm.Name, "VMExtensionDoesNotExist");