+++ src/SDKs/Compute/Compute.Tests/ScenarioTests/VMOperationalTests.cs
@@ -86,7 +86,7 @@ namespace Compute.Tests
                     // Create Storage Account, so that both the VMs can share it
                     var storageAccountOutput = CreateStorageAccount(rg1Name, storageAccountName);
 
-                    VirtualMachine vm1 = CreateVM_NoAsyncTracking(rg1Name, as1Name, storageAccountOutput, imageRef, out inputVM1);
+                    VirtualMachine vm1 = CreateVM(rg1Name, as1Name, storageAccountOutput, imageRef, out inputVM1);
 
                     m_CrpClient.VirtualMachines.Start(rg1Name, vm1.Name);
                     m_CrpClient.VirtualMachines.Redeploy(rg1Name, vm1.Name);
@@ -137,7 +137,7 @@ namespace Compute.Tests
                     // TODO : Provisioning Time-out Issues
                     VirtualMachine inputVM2;
                     string as2Name = as1Name + "b";
-                    VirtualMachine vm2 = CreateVM_NoAsyncTracking(rg1Name, as2Name, storageAccountOutput, imageRef, out inputVM2,
+                    VirtualMachine vm2 = CreateVM(rg1Name, as2Name, storageAccountOutput, imageRef, out inputVM2,
                         vm =>
                         {
                             vm.StorageProfile.ImageReference = null;
@@ -187,7 +187,7 @@ namespace Compute.Tests
                     // Create Storage Account, so that both the VMs can share it
                     var storageAccountOutput = CreateStorageAccount(rg1Name, storageAccountName);
 
-                    VirtualMachine vm1 = CreateVM_NoAsyncTracking(rg1Name, asName, storageAccountOutput, imageRef,
+                    VirtualMachine vm1 = CreateVM(rg1Name, asName, storageAccountOutput, imageRef,
                         out inputVM1);
 
                     var redeployOperationResponse = m_CrpClient.VirtualMachines.BeginRedeployWithHttpMessagesAsync(rg1Name, vm1.Name);
@@ -242,7 +242,7 @@ namespace Compute.Tests
                     // Create Storage Account, so that both the VMs can share it
                     var storageAccountOutput = CreateStorageAccount(rg1Name, storageAccountName);
 
-                    VirtualMachine vm1 = CreateVM_NoAsyncTracking(rg1Name, asName, storageAccountOutput, imageRef,
+                    VirtualMachine vm1 = CreateVM(rg1Name, asName, storageAccountOutput, imageRef,
                         out inputVM1);
 
                     m_CrpClient.VirtualMachines.PerformMaintenance(rg1Name, vm1.Name);