+++ src/SDKs/Compute/Compute.Tests/ScenarioTests/VMScenarioTests.cs
@@ -8,6 +8,7 @@ using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using Xunit;
 
 namespace Compute.Tests
@@ -25,7 +26,6 @@ namespace Compute.Tests
         /// GETVMs in a RG
         /// List VMSizes in a RG
         /// List VMSizes in an AvailabilitySet
-        /// Delete VM
         /// Delete RG
         /// </summary>
         [Fact]
@@ -39,20 +39,33 @@ namespace Compute.Tests
         /// Covers following Operations for managed disks:
         /// Create RG
         /// Create Network Resources
-        /// Create VM
+        /// Create VM with WriteAccelerator enabled OS and Data disk
         /// GET VM Model View
         /// GET VM InstanceView
         /// GETVMs in a RG
         /// List VMSizes in a RG
         /// List VMSizes in an AvailabilitySet
-        /// Delete VM
         /// Delete RG
+        /// 
+        /// To record this test case, you need to run it in region which support XMF VMSizeFamily like eastus2.
         /// </summary>
         [Fact]
         [Trait("Name", "TestVMScenarioOperations_ManagedDisks")]
         public void TestVMScenarioOperations_ManagedDisks()
         {
-            TestVMScenarioOperationsInternal("TestVMScenarioOperations_ManagedDisks", hasManagedDisks: true);
+            string originalTestLocation = Environment.GetEnvironmentVariable("AZURE_VM_TEST_LOCATION");
+            try
+            {
+                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", "eastus2");
+                TestVMScenarioOperationsInternal("TestVMScenarioOperations_ManagedDisks", vmSize: VirtualMachineSizeTypes.StandardM64s, hasManagedDisks: true,
+                    storageAccountType: StorageAccountTypes.PremiumLRS, writeAcceleratorEnabled: true);
+            }
+            finally
+            {
+                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", originalTestLocation);
+            }
+
+            
         }
 
         /// <summary>
@@ -74,13 +87,15 @@ namespace Compute.Tests
             }
         }
 
-        private void TestVMScenarioOperationsInternal(string methodName, bool hasManagedDisks = false, IList<string> zones = null)
+        private void TestVMScenarioOperationsInternal(string methodName, bool hasManagedDisks = false, IList<string> zones = null, string vmSize = VirtualMachineSizeTypes.StandardA0,
+            StorageAccountTypes storageAccountType = StorageAccountTypes.StandardLRS, bool? writeAcceleratorEnabled = null)
         {
             using (MockContext context = MockContext.Start(this.GetType().FullName, methodName))
             {
                 EnsureClientsInitialized(context);
 
                 ImageReference imageRef = GetPlatformVMImage(useWindowsImage: true);
+                const string expectedOSName = "Windows Server 2012 R2 Datacenter", expectedOSVersion = "Microsoft Windows NT 6.3.9600.0", expectedComputerName = ComputerName;
                 // Create resource group
                 var rgName = ComputeManagementTestUtilities.GenerateName(TestPrefix);
                 string storageAccountName = ComputeManagementTestUtilities.GenerateName(TestPrefix);
@@ -90,41 +105,46 @@ namespace Compute.Tests
                 {
                     if (!hasManagedDisks)
                     {
-                        // Create Storage Account, so that both the VMs can share it
-                        var storageAccountOutput = CreateStorageAccount(rgName, storageAccountName);
-
-                        m_CrpClient.VirtualMachines.Delete(rgName, "VMDoesNotExist");
-
-                        m_CrpClient.AvailabilitySets.Delete(rgName, "ASDoesNotExist");
+                        CreateStorageAccount(rgName, storageAccountName);
                     }
 
-                    var vm1 = CreateVM_NoAsyncTracking(rgName, asName, storageAccountName, imageRef, out inputVM, hasManagedDisks: hasManagedDisks, zones: zones);
+                    CreateVM(rgName, asName, storageAccountName, imageRef, out inputVM, hasManagedDisks: hasManagedDisks, vmSize: vmSize, storageAccountType: storageAccountType, 
+                        writeAcceleratorEnabled: writeAcceleratorEnabled, zones: zones);
+                    // NOTE: In record mode, uncomment this line. This is to ensure that there is sufficient time for VMAgent to populate status blob with OS details.
+                    // Thread.Sleep(TimeSpan.FromMinutes(5));
 
                     var getVMWithInstanceViewResponse = m_CrpClient.VirtualMachines.Get(rgName, inputVM.Name, InstanceViewTypes.InstanceView);
                     Assert.True(getVMWithInstanceViewResponse != null, "VM in Get");
-                    ValidateVMInstanceView(inputVM, getVMWithInstanceViewResponse, hasManagedDisks);
+                    ValidateVMInstanceView(inputVM, getVMWithInstanceViewResponse, hasManagedDisks, expectedComputerName, expectedOSName, expectedOSVersion);
 
                     var getVMInstanceViewResponse = m_CrpClient.VirtualMachines.InstanceView(rgName, inputVM.Name);
                     Assert.True(getVMInstanceViewResponse != null, "VM in InstanceView");
-                    ValidateVMInstanceView(inputVM, getVMInstanceViewResponse, hasManagedDisks);
+                    ValidateVMInstanceView(inputVM, getVMInstanceViewResponse, hasManagedDisks, expectedComputerName, expectedOSName, expectedOSVersion);
 
                     bool hasUserDefinedAS = zones == null;
 
                     var listResponse = m_CrpClient.VirtualMachines.List(rgName);
                     ValidateVM(inputVM, listResponse.FirstOrDefault(x => x.Name == inputVM.Name),
-                        Helpers.GetVMReferenceId(m_subId, rgName, inputVM.Name), hasManagedDisks, hasUserDefinedAS);
+                        Helpers.GetVMReferenceId(m_subId, rgName, inputVM.Name), hasManagedDisks, hasUserDefinedAS, writeAcceleratorEnabled);
 
                     var listVMSizesResponse = m_CrpClient.VirtualMachines.ListAvailableSizes(rgName, inputVM.Name);
-                    Helpers.ValidateVirtualMachineSizeListResponse(listVMSizesResponse, hasAZ: zones != null);
+                    Helpers.ValidateVirtualMachineSizeListResponse(listVMSizesResponse, hasAZ: zones != null, writeAcceleratorEnabled: writeAcceleratorEnabled);
 
                     listVMSizesResponse = m_CrpClient.AvailabilitySets.ListAvailableSizes(rgName, asName);
-                    Helpers.ValidateVirtualMachineSizeListResponse(listVMSizesResponse, hasAZ: zones != null);
-
-                    m_CrpClient.VirtualMachines.Delete(rgName, inputVM.Name);
+                    Helpers.ValidateVirtualMachineSizeListResponse(listVMSizesResponse, hasAZ: zones != null, writeAcceleratorEnabled: writeAcceleratorEnabled);
                 }
                 finally
                 {
-                    m_ResourcesClient.ResourceGroups.Delete(rgName);
+                    // Fire and forget. No need to wait for RG deletion completion
+                    try
+                    {
+                        m_ResourcesClient.ResourceGroups.BeginDelete(rgName);
+                    }
+                    catch (Exception e)
+                    {
+                        // Swallow this exception so that the original exception is thrown
+                        Console.WriteLine(e);
+                    }
                 }
             }
         }