+++ src/SDKs/Search/Build-SearchPackage.ps1
@@ -1,2 +1,0 @@
-$scriptDir = Split-Path $MyInvocation.MyCommand.Path -Parent
-msbuild "$scriptDir\..\..\build.proj" /t:"build;package" /p:scope=Search