+++ src/SDKs/DataFactory/Management.DataFactory/changelog.md
@@ -1,5 +1,30 @@
 # Changelog for the Azure Data Factory V2 .NET SDK
 
+## Version 0.6.0-preview
+
+### Feature Additions
+  * SecureOutput and BatchCount
+  * Filter Activity
+  * Linked Service Parameters
+  * FilterActivity
+
+## Version 0.5.0-preview
+
+### Feature Additions
+  * Enable AAD auth via service principal and management service identity for Azure SQL DB/DW linked service types
+  * Support integration runtime sharing across subscription and data factory
+  * Enable Azure Key Vault for all compute linked service
+  * Add SAP ECC Source
+  * GoogleBigQuery support clientId and clientSecret for UserAuthentication
+  * Add LinkedService, Dataset, CopySource for Vertica and Netezza
+
+## Version 0.4.0-preview
+
+### Feature Additions
+  * Add readBehavior to Salesforce Source
+  * Enable Azure Key Vault support for all data store linked services
+  * Add license type property to Azure SSIS integration runtime
+
 ## Version 0.3.0-preview
 
 ### Feature Additions